requirements.txt
# Core dependencies
numpy==1.23.5
pandas==1.5.3

# MQTT for real-time sensor communication
paho-mqtt==1.6.1

# Flask + Dash for the Web Dashboard
flask==2.2.5
dash==2.15.0
dash-bootstrap-components==1.4.1

# Data Visualization
plotly==5.11.0

# Data generation
timesynth==0.2.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Environment variable management
python-dotenv==1.0.0
